move enums.py
